requirements.txt
annotated-types==0.7.0
boto3==1.36.17
botocore==1.36.17
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.67.0rc1
grpcio==1.70.0
grpcio-status==1.70.0
httplib2==0.22.0
idna==3.10
jmespath==1.0.1
pinecone==6.0.1
pinecone-plugin-interface==0.0.7
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
s3transfer==0.11.2
six==1.17.0
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
